Moved tracked methods out of tensor (#11921)
* #0: moved tracked methods out of tensor

* #0: updated cmake list, removed copypaste

* #0: sync cleanup

* #0: fixed includes
