+++ ttnn/cpp/ttnn/tensor/CMakeLists.txt
@@ -2,6 +2,7 @@
 set(TENSOR_SRCS
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor_impl.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor.cpp
+    ${CMAKE_CURRENT_SOURCE_DIR}/tensor_ops.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/types.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/tensor_utils.cpp
     ${CMAKE_CURRENT_SOURCE_DIR}/serialization.cpp
